Moves tests to dronekit.test folder.
