+++ .travis.yml
@@ -11,9 +11,9 @@ install:
   - 'sudo pip install -r requirements.txt -UI'
 
 script:
-  - '[ "${TRAVIS_PULL_REQUEST}" != "false" ] || nosetests tests/web'
-  - 'nosetests tests/unit'
-  - 'nosetests tests/sitl'
+  - '[ "${TRAVIS_PULL_REQUEST}" != "false" ] || nosetests dronekit.test.web'
+  - 'nosetests dronekit.test.unit'
+  - 'nosetests dronekit.test.sitl'
 
 git:
   depth: 10
