+++ dronekit/test/sitl/test_goto.py
@@ -9,7 +9,7 @@ Full documentation is provided at http://python.dronekit.io/examples/simple_goto
 
 import time
 from dronekit import connect, VehicleMode, LocationGlobal
-from dronekit.tools import with_sitl
+from dronekit.test import with_sitl
 from nose.tools import assert_equals
 
 @with_sitl
