+++ dronekit/test/web/cloud_client_test.py
