+++ dronekit/util.py
@@ -1,0 +1,5 @@
+from __future__ import print_function
+import sys
+
+def errprinter(*args):
+    print(*args, file=sys.stderr)
